backend/pyproject.toml
[project]
name = "gov-data-api"
version = "0.1.0"
description = "FastAPI REST API for government data pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.0",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "minio>=7.2",
    "python-dotenv>=1.0",
    "anthropic>=0.40",
    "sentence-transformers>=3.0",
    "weaviate-client>=4.9",
    "neo4j>=5.0",
    "firecrawl-py>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
